AOCL-LibMem: Fix for SEGFAULT with Tunable binary
In the Tunables library, the constructor calls strcmp function
for parsing the environment variables. This leads to segfault
as the strcmp is implemented inside the library and marked as a
weak symbol to be resolved during the init phase.

Fixed operator precedence issue to avoid segfaults in memmove.

Also fixed the incorrect resolution of the symbol function pointer
for unknown tunable options.

AMD-Internal: [CPUPL-6164]
Change-Id: I725cb22391e534471e5adc9ee3db40b761b5f126