AOCL-LibMem: Benchmark enhancement for logging system information
An optional system information collection step is added to the pre-benchmark
set-up. It gathers details about CPU, memory, and frequency settings,
and stores the information in the system_info.txt in the result dir.

To enable logging system information, '-sys' option needs to be provided
in the benchmark command.

AMD-Internal: [CPUPL-6536]
Change-Id: Ifdde3758ddb70d04372c22d10d8a6265efe0a5d6