AOCL-LibMem: Prepare buffers correctly for memcmp tinybenchmark(TBM)
This commit fixes an issue in the tinymembench patch where the buffers
for the memcmp benchmark were not correctly prepared.
It ensures that for the memcmp benchmark, the source and destination
buffers contain identical data by performing a memcpy before the test.
This allows for an accurate performance measurement of comparing equal
memory regions.

AMD-Internal: [CPUPL-7170]